Move rpc handler validation utils to validation module
