+++ newsfragments/174.misc.rst
@@ -1,0 +1,1 @@
+Relocate rpc handler validation utils to validation module.
